pyproject.toml
[project]
name = "clustering"
version = "0.1.0"
description = "3D point cloud semantic segmentation using OpenCLIP + VGGT"
requires-python = ">=3.10,<3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "pillow>=12.2.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=5.5.0",
    "vggt"
]

[tool.uv.sources]
vggt = { path = "../VGGT", editable = true }